Fix subproject to pass test 'pull-after-push'
v0.1.1 refined parsing of the push markers of 'their branch' (that is
the subproject's source branch). However the change did not consider,
that the variable prefixes MY_ and THEIR_ are relative to 'my branch'
(that is the branch containing the subproject). Hence, when parsing
their branch's history, the marker argument -Xtheir-prefix needs to be
matched against MY_PREFIX.

Besides above fix, this commit further improves the regular expression
used for parsing their branch's push markers: Positive look ahead is
used to match the push marker arguments -Xtheir-prefix and
-Xtheir-branch independent of the order in which the arguments are
present.