Identify push commits based on subproject id
A subproject id has been introduced by this commit. The id is generated
during init - or for subprojects created by an older version of this
tool during the next operation. Later on, the id is used to recognize
push commits when parsing their history.

The current implementation is used as fallback, so there shouldn't be
any trouble when using the new subproject utility with subprojects that
have initially been created with an older version. An id is silently
added to these repositories.

Gitlab issue: closes #13